# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Exam Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results/{attemptId}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get result by attempt ID
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultDetail"
  /results/{attemptId}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [results]
      summary: Get leaderboard for an exam
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # PREVIOUS YEAR PAPERS
...
components:
  schemas:
...
    ResultDetail:
      type: object
      required: [id, attemptId, userId, totalScore, maxScore, correctCount, wrongCount, skippedCount, accuracy, createdAt, answers]
      properties:
        id:
          type: integer
        attemptId:
          type: integer
        userId:
          type: integer
        totalScore:
          type: number
        maxScore:
          type: number
        correctCount:
          type: integer
        wrongCount:
          type: integer
        skippedCount:
          type: integer
        accuracy:
          type: number
        timeTaken:
          type: ["integer", "null"]
        rank:
          type: ["integer", "null"]
        percentile:
          type: ["number", "null"]
        examTitle:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        sectionAnalysis:
          type: array
          items:
            $ref: "#/components/schemas/SectionAnalysis"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerResult"
    SectionAnalysis:
      type: object
      required: [sectionId, sectionName, correct, wrong, skipped, score]
      properties:
        sectionId:
          type: integer
        sectionName:
          type: string
        correct:
          type: integer
        wrong:
          type: integer
        skipped:
          type: integer
        score:
          type: number
        timeTaken:
          type: ["integer", "null"]
    AnswerResult:
      type: object
      required: [questionId, questionText, selectedOption, correctAnswer, isCorrect, marks]
      properties:
        questionId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        selectedOption:
          type: ["string", "null"]
        correctAnswer:
          type: string
        isCorrect:
          type: boolean
        markedForReview:
          type: boolean
        marks:
          type: number
        explanation:
          type: ["string", "null"]
    LeaderboardEntry:
      type: object
      required: [rank, userId, userName, totalScore, accuracy, timeTaken]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        totalScore:
          type: number
        accuracy:
          type: number
        timeTaken:
          type: integer

    # PAPERS
...